Use BIGINT instead of INT for valid schema

+++ spring-boot-datajpatest/src/main/resources/db/migration/V1__create_order_and_payment.sql
@@ -2,15 +2,15 @@ CREATE SEQUENCE hibernate_sequence
     START 1 INCREMENT 1;
 
 CREATE TABLE orders(
-    id INT NOT NULL PRIMARY KEY,
+    id BIGINT NOT NULL PRIMARY KEY,
     date TIMESTAMP NOT NULL,
     amount NUMERIC NOT NULL,
     paid BOOLEAN NOT NULL
 );
 
 CREATE TABLE payment(
-    id INT NOT NULL PRIMARY KEY,
-    order_id INT NOT NULL UNIQUE,
+    id BIGINT NOT NULL PRIMARY KEY,
+    order_id BIGINT NOT NULL UNIQUE,
     credit_card_number VARCHAR(16) NOT NULL,
     CONSTRAINT fk_order FOREIGN KEY(order_id) REFERENCES orders(id)
 );
